+++ Calculator/Functions.h
@@ -1,0 +1,9 @@
+#pragma once
+
+double get_value(string s);
+void set_value(string s, double d);
+bool is_declared(string var);
+double define_name(string var, double val);
+void calculate();
+
+
