+++ Calculator/Variable.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include <string>
 
 using std::string;
